requirements.txt
quart==0.19.4
python-dotenv==1.0.0
discord.py==2.3.2
requests==2.31.0
gunicorn==21.2.0 